finished login design
