--- src/bin/stack/app/src/Index.svelte
@@ -1,8 +1,0 @@
-<script>
-</script>
-<main>
-
-</main>
-<style>
-
-</style>
